apollo-server: factor out gui checking code
